crm: fix the two GCC diagnostics the first cast error was hiding
The ladder leg builds gcc-debug, and it used to die on the -Wuseless-cast in
tests/test_extension_bag_spike.cpp before ninja ever reached examples/crm/.
With that fixed the build gets further and finds two more, both GCC-only --
which is why the clang-coverage verification of the previous commit did not
catch them:

- crm_schemas.hpp: a second useless cast of `size()`, identical in shape to
  the one already fixed in the spike test. Same fix.
- custom_field_dto.hpp: -Wshadow on `CustomFieldType::Money`, which matches
  the `crm::Money` alias it deliberately names.

The shadow one is suppressed locally rather than renamed. GCC reports an
enumerator matching a namespace-scope name even for a *scoped* enum, where
the collision cannot occur: `CustomFieldType::Money` is never found by
unqualified lookup. Renaming either side would cost more than it buys --
the enumerator's spelling is the wire representation (`glz::enumerate`
serialises by name, so a rename changes stored and served data), and the
alias is used across the rung's DTOs, entity and tests. The suppression is
scoped to the enum and carries its reasoning, in the same shape
`morph::OpaqueIdGenerator` already suppresses -Wuseless-cast. Clang does not
warn here.

Verified in the configuration that failed this time (gcc-debug,
MORPH_BUILD_LADDER=ON, MORPH_LADDER_RUNGS=all): all 388 targets build with
zero errors, and `ctest -L ladder -LE stress` is 920/920 green with crm's
own 174 among them.

+++ examples/crm/include/crm/gui/crm_schemas.hpp
@@ -78,7 +78,12 @@ namespace crm::gui {
 
     std::uint64_t nextOrder = 0;
     if (dom.contains("properties") && dom["properties"].holds<glz::generic_u64::object_t>()) {
-        nextOrder = static_cast<std::uint64_t>(dom["properties"].get<glz::generic_u64::object_t>().size());
+        // No cast: `size()` returns `std::size_t`, which is the same type as
+        // `std::uint64_t` on the 64-bit targets this header is built for, so an
+        // explicit cast is a no-op GCC rejects under -Wuseless-cast. Were
+        // `std::size_t` ever narrower, the assignment widens implicitly and
+        // safely -- unsigned to wider unsigned preserves the value.
+        nextOrder = dom["properties"].get<glz::generic_u64::object_t>().size();
     }
 
     for (const auto& field : fields) {
